Support testing network-activation
As part of this we no longer just simulate plugin activation and
uninstallation, but actually use `uninstall_plugin()` and
`activate_plugin()`. This makes the uninstall and activation almost
100% realistic.

At the same time, we’ve also improved realism by deactivating the
plugin after usage simulation, since in real life a plugin can only be
uninstalled after it is deactivated.

As a consequence of all of this, we now expect the plugin folder to be
physically installed on the site being used during the tests, that is,
the plugin’s folder needs to be either copied or symlinked to the
plugin directory. We also expect `$this->plugin_file` to be relative to
the plugin directory now.

To run the tests with the plugin network active instead of just
activated on a single site on the network, you need to set
`$this->network_active` to `true` within the test case class. Perhaps
we’ll also add support for setting this via a command line flag or
something later on.

Fixes #2